app_new.py
import streamlit as st
import os
from dotenv import load_dotenv
import openai
from supabase import create_client
import pinecone
from llama_index.core import VectorStoreIndex, ServiceContext
from llama_index.vector_stores.pinecone import PineconeVectorStore
from llama_index.embeddings.openai import OpenAIEmbedding
from llama_index.llms.openai import OpenAI
import fitz  # PyMuPDF
import tempfile
import uuid

# Load environment variables
load_dotenv()

# Configure page
st.set_page_config(
    page_title="Coursera Study Buddy",
    page_icon="📚",
    layout="wide",
)

# Initialize API clients
# NOTE: API initialization code will go here when environment variables are properly set up

# App UI
st.title("📚 Coursera Study Buddy")
st.subheader("Upload your course materials and ask questions")

# Sidebar
with st.sidebar:
    st.header("About")
    st.write("This app helps you study Coursera materials by allowing you to upload PDFs and ask questions about the content.")
    
    st.header("Settings")
    api_key = st.text_input("OpenAI API Key (optional)", type="password", help="Enter your OpenAI API key if not set in environment variables")

# Main content
tab1, tab2 = st.tabs(["Upload", "Chat"])

with tab1:
    st.header("Upload Course Materials")
    uploaded_file = st.file_uploader("Choose a PDF file", type="pdf")
    
    if uploaded_file is not None:
        st.success(f"File '{uploaded_file.name}' uploaded successfully!")
        
        if st.button("Process PDF"):
            with st.spinner("Processing your PDF..."):
                # Here we would process the PDF and store it in the vector database
                st.session_state.file_processed = True
                st.success("PDF processed successfully! You can now chat with your document.")

with tab2:
    st.header("Chat with Your Course Materials")
    
    if "messages" not in st.session_state:
        st.session_state.messages = []
    
    # Display chat messages
    for message in st.session_state.messages:
        with st.chat_message(message["role"]):
            st.write(message["content"])
    
    # Chat input
    if prompt := st.chat_input("Ask a question about your course materials"):
        # Add user message to chat history
        st.session_state.messages.append({"role": "user", "content": prompt})
        
        # Display user message in chat message container
        with st.chat_message("user"):
            st.write(prompt)
        
        # Display assistant response in chat message container
        with st.chat_message("assistant"):
            if not st.session_state.get("file_processed", False):
                response = "Please upload and process a PDF file first before asking questions."
            else:
                response = f"This is a placeholder response to your question: '{prompt}'. In the full implementation, this would query the vector database."
            
            st.write(response)
            
        # Add assistant response to chat history
        st.session_state.messages.append({"role": "assistant", "content": response})
