requirements.txt
streamlit>=1.27.2
openai>=1.3.0
python-dotenv>=1.0.0
supabase>=1.0.3
pinecone-client>=2.2.4
llama-index>=0.9.48
llama-index-vector-stores-pinecone>=0.1.4
llama-index-embeddings-openai>=0.1.5
llama-index-llms-openai>=0.1.5
pymupdf>=1.23.3 